requirements.txt
attrs==20.3.0
black==20.8b1
flake8==3.8.4
ipdb==0.13.4
isort==5.6.4
mypy==0.790
pytest==6.1.2
